improve implementation
